refactor: 파일 공통화
