+++ apps/web/src/app/[locale]/mypage/my-pet/(evolution)/EvolutionPersona.tsx
@@ -1,36 +1,34 @@
 'use client';
 
-import { memo, useState } from 'react';
+import { useState } from 'react';
 import { useTranslations } from 'next-intl';
-import { css, cx } from '_panda/css';
+import { css } from '_panda/css';
 import type { MergePersonaLevelResponse, Persona } from '@gitanimals/api';
 import { useMergePersonaLevelByToken, userQueries } from '@gitanimals/react-query';
-import { Button, Dialog, LevelBanner } from '@gitanimals/ui-panda';
-import { BannerSkeletonList } from '@gitanimals/ui-panda/src/components/Banner/Banner';
-import { wrap } from '@suspensive/react';
-import { useQueryClient, useSuspenseQuery } from '@tanstack/react-query';
+import { Button, Dialog } from '@gitanimals/ui-panda';
+import { useQueryClient } from '@tanstack/react-query';
 
-import { customScrollStyle } from '@/styles/scrollStyle';
-import { useClientSession, useClientUser } from '@/utils/clientAuth';
-import { getPersonaImage } from '@/utils/image';
+import { useClientSession } from '@/utils/clientAuth';
 
-import { MergePreview } from './MergePreview';
+import { SelectPersonaList } from '../../PersonaList';
+import { SpinningLoader } from '../_components/SpinningLoader';
+
+import { EvolutionPreview } from './EvolutionPreview';
 import { MergeResultModal } from './MergeResult';
 
-interface MergePersonaProps {
+interface EvolutionPersonaProps {
   isOpen: boolean;
   onClose: () => void;
   targetPersona: Persona;
 }
 
-export function EvolutionPersona({ isOpen, onClose, targetPersona: initTargetPersona }: MergePersonaProps) {
+export function EvolutionPersona({ isOpen, onClose, targetPersona }: EvolutionPersonaProps) {
   const queryClient = useQueryClient();
   const t = useTranslations('Mypage.Merge');
 
   const [resultData, setResultData] = useState<MergePersonaLevelResponse | null>(null);
 
   const [materialPersona, setMaterialPersona] = useState<Persona | null>(null);
-  const [targetPersona, setTargetPersona] = useState<Persona>(initTargetPersona);
 
   const session = useClientSession();
   const token = session.data?.user.accessToken as string;
@@ -39,9 +37,6 @@ export function EvolutionPersona({ isOpen, onClose, targetPersona: initTargetPer
 
   const { mutate: mergePersonaLevel, isPending: isMerging } = useMergePersonaLevelByToken(token, {
     onSuccess: (data) => {
-      // setMaterialPersona(null);
-      // setResultData(data);
-      // setTargetPersona(data as Persona); // TODO: 타입 확인
       queryClient.invalidateQueries({ queryKey: userQueries.allPersonasKey() });
     },
   });
@@ -69,7 +64,7 @@ export function EvolutionPersona({ isOpen, onClose, targetPersona: initTargetPer
     <Dialog open={isOpen} onOpenChange={onClose}>
       <Dialog.Content size="large">
         <Dialog.Title className={headingStyle}>Merge to Level Up</Dialog.Title>
-        <MergePreview targetPersona={targetPersona} materialPersona={materialPersona} />
+        <EvolutionPreview targetPersona={targetPersona} materialPersona={materialPersona} />
 
         <SelectPersonaList selectPersona={selectPersona} onSelectPersona={onSelectPersona} />
 
@@ -95,124 +90,3 @@ export function EvolutionPersona({ isOpen, onClose, targetPersona: initTargetPer
 
 const headingStyle = css({ marginTop: '40px' });
 const footerStyle = css({ display: 'flex', justifyContent: 'center', gap: '12px' });
-
-interface SelectPersonaListProps {
-  selectPersona: string[];
-  onSelectPersona: (persona: Persona) => void;
-  initSelectPersonas?: (list: Persona[]) => void;
-  loadingPersona?: string[];
-}
-
-const SelectPersonaList = wrap
-  .ErrorBoundary({ fallback: <div>error</div> })
-  .Suspense({ fallback: <BannerSkeletonList length={6} size="small" /> })
-  .on(function SelectPersonaList({ selectPersona, onSelectPersona }: SelectPersonaListProps) {
-    const { name } = useClientUser();
-    const { data } = useSuspenseQuery(userQueries.allPersonasOptions(name));
-    const t = useTranslations('Mypage.Merge');
-
-    // TODO: 정렬
-    return (
-      <section className={sectionStyle}>
-        <div className={listSectionTitleStyle}>
-          <p>{t('please-choose-pet')}</p>
-        </div>
-        <div className={flexOverflowStyle}>
-          {data.personas.map((persona) => (
-            <MemoizedPersonaItem
-              key={persona.id}
-              persona={persona}
-              isSelected={selectPersona.includes(persona.id)}
-              onClick={() => onSelectPersona(persona)}
-            />
-          ))}
-        </div>
-      </section>
-    );
-  });
-
-const sectionStyle = css({
-  height: '100%',
-  maxHeight: '100%',
-  minHeight: '0',
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '16px',
-});
-
-const listSectionTitleStyle = css({
-  textStyle: 'glyph16.regular',
-  color: 'white.white_50',
-  display: 'flex',
-  justifyContent: 'space-between',
-});
-
-const flexOverflowStyle = cx(
-  css({
-    display: 'flex',
-    overflowY: 'auto',
-    overflowX: 'hidden',
-    width: '100%',
-    gap: '4px',
-    height: '100%',
-    minHeight: '0',
-    flexWrap: 'wrap',
-    maxHeight: 'calc(100% - 24px)',
-  }),
-  customScrollStyle,
-);
-
-interface PersonaItemProps {
-  persona: Persona;
-  isSelected: boolean;
-  onClick: () => void;
-}
-
-function PersonaItem({ persona, isSelected, onClick }: PersonaItemProps) {
-  return (
-    <button onClick={onClick} className={css({ outline: 'none', bg: 'transparent' })}>
-      <LevelBanner
-        image={getPersonaImage(persona.type)}
-        status={isSelected ? 'selected' : 'default'}
-        level={Number(persona.level)}
-        size="small"
-      />
-    </button>
-  );
-}
-
-const MemoizedPersonaItem = memo(PersonaItem, (prev, next) => {
-  return prev.isSelected === next.isSelected && prev.persona.level === next.persona.level;
-});
-
-const SpinningLoader = () => {
-  return (
-    <div className={spinningLoaderContainerStyle}>
-      <div className={spinningLoaderStyle} />
-    </div>
-  );
-};
-
-const spinningLoaderContainerStyle = css({
-  position: 'absolute',
-  top: '0',
-  left: '0',
-  right: '0',
-  bottom: '0',
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  width: '100%',
-  zIndex: 'floating',
-  height: '100%',
-  background: 'rgba(0, 0, 0, 0.5)',
-});
-
-const spinningLoaderStyle = css({
-  width: '64px',
-  height: '64px',
-  border: '4px solid transparent',
-  borderTop: '4px solid #fff',
-  borderRadius: '50%',
-  animation: 'animateSpin 1s linear infinite',
-});
