+++ packages/api/src/render/index.ts
@@ -7,3 +7,4 @@ export * from './isPressStar';
 export * from './getMyBackground';
 export * from './changeMyBackground';
 export * from './mergePersonaLevel';
+export * from './evolution';
