+++ packages/api/src/render/schema.ts
@@ -8,6 +8,7 @@ export const PersonaSchema = z.object({
   visible: z.boolean(),
   dropRate: z.string(),
   grade: PersonaGradeSchema,
+  isEvolution: z.boolean().optional(),
 });
 
 export const PersonaInfoSchema = PersonaSchema.pick({
